Mock fidelity: GPoint/GSize match the SDK's int16_t
Host arithmetic in int can't reproduce device int16_t truncation; the mock's
geometry types now match the SDK widths. Template hygiene — today's faces
can't overflow, a fork's might.

+++ test/pebble.h
@@ -19,12 +19,12 @@ typedef struct DictionaryIterator DictionaryIterator;
 typedef void AppTimer;
 
 typedef struct {
-  int x;
-  int y;
+  int16_t x;
+  int16_t y;
 } GPoint;
 typedef struct {
-  int w;
-  int h;
+  int16_t w;
+  int16_t h;
 } GSize;
 typedef struct {
   GPoint origin;
